Moves boost json and vector support to utility.
