+++ test/test_cereal_boost_container.cpp
@@ -3,25 +3,24 @@
 //
 // SPDX-License-Identifier: MIT
 
-
 #undef NDEBUG
 
 #include <vector>
 
 #include <boost/container/vector.hpp>
 
-#include <ygm/detail/assert.hpp>
 #include <ygm/detail/byte_vector.hpp>
-#include <ygm/detail/cereal_boost_container.hpp>
 #include <ygm/detail/ygm_cereal_archive.hpp>
+#include <ygm/utility/assert.hpp>
+#include <ygm/utility/boost_vector.hpp>
 
 int main() {
   // Currently, we only support boost::container::vector
 
   boost::container::vector<int> original_value = {1, 2, 3, 4, 5};
-  ygm::detail::byte_vector cereal_buffer;
+  ygm::detail::byte_vector      cereal_buffer;
   {
-    cereal::YGMOutputArchive      archive(cereal_buffer);
+    cereal::YGMOutputArchive archive(cereal_buffer);
     archive(original_value);
   }
